fdt/rtems-fdt-shell.c: Correct comparisons of different signedness
The GCC warning -Wsign-compare flagged these instances of comparing
signed and unsigned types. A common pair was size_t and int.

+++ cpukit/libmisc/rtems-fdt/rtems-fdt-shell.c
@@ -189,7 +189,7 @@ rtems_fdt_shell_ls (int argc, char *argv[])
   size_t path_len = 0;
   int    total_entries = 0;
   int    num_entries = 0;
-  int    max_name_len = 0;
+  size_t max_name_len = 0;
   int    name_offset = 0;
   int    i = 0;
